Update PRIVACY.md
Clarify that Firebase Analytics is in the app but it's just collecting default analytics. I don't see anything special being done with it or the username. Clarify what photos/videos collection is done (i.e. none, just Imgur uploading) because the data collection policy on the Play Store doesn't have an ability to clarify how exactly photos/videos are accessed.